feat: refine evaluation metric chart layout and controls
